requirements/runtime.txt
appdirs==1.4.0
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.17.1
channels==1.0.3
constantly==15.1.0
daphne==1.0.3
Django==1.10.5
django-filter==1.0.1
django-model-utils==2.6.1
django-object-actions==0.9.0
djangorestframework==3.5.4
incremental==16.10.1
Jinja2==2.9.5
MarkupSafe==0.23
msgpack-python==0.4.8
packaging==16.8
panavatar==0.3.3
PyMySQL==0.7.10
pyparsing==2.1.10
redis==2.10.5
rules==1.2
six==1.10.0
Twisted==16.5.0
txaio==2.6.1
uWSGI==2.0.14
xlrd==1.0.0
zope.interface==4.3.3
